Broke down frontend into smaller components.
