+++ backend/uploads/211a09fe-7263-4312-b6f9-cbe5d561310d-celsius-ferenheit.c
@@ -1,0 +1,15 @@
+#include <stdio.h>
+
+int main()
+{
+
+    float c;
+
+    printf("Enter C:");
+    scanf("%f", &c);
+
+    float f;
+    f = (9 * c) / 5 + 32;
+
+    printf("Farenheit: %f", f);
+}
