+++ backend/uploads/440abec3-77ea-4144-9667-5a755c0144fa-celsius-ferenheit.c
@@ -1,0 +1,15 @@
+#include <stdio.h>
+
+int main()
+{
+
+    float c;
+
+    printf("Enter C:");
+    scanf("%f", &c);
+
+    float f;
+    f = (9 * c) / 5 + 32;
+
+    printf("Farenheit: %f", f);
+}
