+++ backend/uploads/51f4dbd5-94cf-46fe-9799-e0f1978152f0-celsius-ferenheit.c
@@ -1,0 +1,15 @@
+#include <stdio.h>
+
+int main()
+{
+
+    float c;
+
+    printf("Enter C:");
+    scanf("%f", &c);
+
+    float f;
+    f = (9 * c) / 5 + 32;
+
+    printf("Farenheit: %f", f);
+}
